refactor: move home page rendering to pages/home.rs
